add isDependent(), update Deploy Block filtering
This adds in the isDependent() function, which will check if the pull
request base ref (branch that it is based off of) is on the main/master
branch or not.

This function returns a bool which will return 'true' if the pull
request's base branch is not the main/master branch (making it
dependent) or 'false' if the pull request is independent, and won't need
to wait for another pull request to merge first.

This commit also updates the Deploy Block column filtering by including
dependent pull requests that have met deploy requirements.

Note: The base ref (base branch) will automatically switch to master
once the parent pull request has been merged, which will then update
this flag state.