add new pull card
This adds in a new pull card named 'PullCardWithLinesChanged'. This new card
will copy everything over from PullCard with the addition of a Code
Block, HStack, and Text chakra component.

The purpose of this card is to show the pull request additions &
deletions for CR purposes (# of lines changed for each pull request)