move connection state + open pull count
This moves the UI component that displays the current connection status
along with the # of open pulls over to the right-hand side of the navigation
bar (before the Search bar).

I introduced a <Text> component from Chakra and leveraged a HStack to
align this section. I also separated the text of the "open:" and the
number of open pulls to make sure that there isn't a horizontal shift in
the user interface when the number of open pulls changes.