DB: fix DB connection and query format
Ooof, see the second change.

First: The .query().then() was just changed to make use of the new
promise interface instead of promisifying the old callback interface.
It wasn't the problem.

Second: Add support for utf8mb3 to the library (we'll open a pull to add
it upstream). See the code comment.