frontend: use index.html template for pull-card-demo
Turns out, this was generating pull-card-demo.html which isn't' what's
used when accessing /pull-card-demo (that is dynamically generated using
webpack? and doesn't include the body classes we added for chakra).

Now /pull-card-demo should load the generated index.html and thus have
the same look and feel as the main page.